--- .vercelignore
@@ -1,2 +1,0 @@
-/target
-rust.sh
