--- api/_utils/mod.rs
@@ -1,9 +1,0 @@
-pub mod assigner;
-pub mod builder;
-pub mod data;
-pub mod dna;
-pub mod error;
-pub mod genetic_circuit;
-pub mod lexer;
-pub mod logic_circuit;
-pub mod parser;
