+++ src/genetic_circuit/component.rs
