+++ src/logic_circuit/input.rs
@@ -1,4 +1,4 @@
-use crate::_utils::{data, genetic_circuit};
+use crate::{genetic_circuit, utils::data};
 use data::get_data;
 use genetic_circuit::{Component, Signal};
 use std::collections::hash_map::HashMap;
