--- vercel.json
@@ -1,9 +1,0 @@
-{
-	"version": 2,
-	"builds": [
-		{
-			"src": "api/compile.rs",
-			"use": "@genhub/vercel-rust"
-		}
-	]
-}
